+++ src/commands/mod.rs
@@ -2,3 +2,4 @@ pub mod clear_cmd;
 pub mod config_cmd;
 pub mod gen_cmd;
 pub mod new_cmd;
+pub mod verify_cmd;